requirements.txt
Flask==2.3.2
gunicorn==21.2.0
line-bot-sdk==3.5.0
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.2